local-streamlit-app/requirements.txt
cohere==4.37
llama-index==0.10.17
llama-index-agent-openai==0.1.5
llama-index-llms-openai==0.1.7
llama-index-readers-youtube-transcript==0.1.4
llama-index-vector-stores-deeplake==0.1.2
openai==1.3.8 
python==3.1.2
python-dotenv==1.0.1
stremlit==1.32.0